refactor: backend maintenance hook and imports
