+++ apps/main/src/routes/GlobalLayout.tsx
@@ -1,4 +1,3 @@
-import { BackendMaintenanceModal } from 'curve-ui-kit/src/features/maintenance/BackendMaintenanceModal'
 import { type ReactNode } from 'react'
 import { networks as daoNetworks } from '@/dao/networks'
 import { useDexAppStats, useDexRoutes } from '@/dex/hooks/useDexAppStats'
@@ -82,7 +81,6 @@ export const GlobalLayout = <TId extends string, TChainId extends number>({
     >
       <ErrorBoundary title={t`Page error`}>{children}</ErrorBoundary>
     </Box>
-    <BackendMaintenanceModal />
     <Footer appName={currentApp} networkId={network.id} />
   </Stack>
 )
